edit btn and weight fixed

+++ app/src/main/java/com/example/gymproject/models/PartialCustomExercise.java
@@ -3,11 +3,11 @@ package com.example.gymproject.models;
 public class PartialCustomExercise {
     private int sets;
     private int reps;
-    private int weight;
+    private double weight;
     private int rest;
     private String other;
 
-    public PartialCustomExercise(int sets, int reps, int weight, int rest, String other) {
+    public PartialCustomExercise(int sets, int reps, double weight, int rest, String other) {
         this.sets = sets;
         this.reps = reps;
         this.weight = weight;
@@ -17,7 +17,7 @@ public class PartialCustomExercise {
 
     public int getSets() { return sets; }
     public int getReps() { return reps; }
-    public int getWeight() { return weight; }
+    public double getWeight() { return weight; }
     public int getRest() { return rest; }
     public String getOther() { return other; }
 
@@ -27,7 +27,7 @@ public class PartialCustomExercise {
     public void setReps(int reps) {
         this.reps = reps;
     }
-    public void setWeight(int weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
     public void setRest(int rest) {
